Sample's time changed from int to long. SQLData is ready for testing. Also removed app.iml from the git repository to fix stuff.

+++ app/src/main/java/ubw6/com/georeport/Sample.java
@@ -18,7 +18,7 @@ public class Sample {
     public final double myLat;
     public final double myHeading;
     public final double mySpeed;
-    public final int myTime;
+    public final long myTime;
     public final String myUID;
     public final int myPID;
 
@@ -31,7 +31,8 @@ public class Sample {
      * @param theTime Timestamp of the sample (Unix Timestamp format)
      * @param theUID User ID
      */
-    public Sample(double theLon, double theLat, double theHeading, double theSpeed, int theTime, String theUID) {
+    public Sample(double theLon, double theLat, double theHeading,
+                  double theSpeed, long theTime, String theUID) {
         this(theLon, theLat, theHeading, theSpeed, theTime, theUID, -1);
     }
 
@@ -45,7 +46,8 @@ public class Sample {
      * @param theUID User ID
      * @param thePID ID of sample from local database
      */
-    public Sample(double theLon, double theLat, double theHeading, double theSpeed, int theTime, String theUID, int thePID) {
+    public Sample(double theLon, double theLat, double theHeading, double theSpeed,
+                  long theTime, String theUID, int thePID) {
         this.myPID = thePID;
         this.myLon = theLon;
         this.myLat = theLat;
